refactor(nicegui): move filepicker to utils
