+++ anvil/ansible/group_vars/all/container.yml
@@ -1,0 +1,3 @@
+---
+forge_container_file: "Containerfile"
+forge_container_format: "oci"
