+++ anvil/ansible/playbooks/project_build.yml
@@ -14,8 +14,8 @@
         tag: latest
         path: "{{ forge_git_repository_destination }}"
         build:
-          file: Containerfile
-          format: oci
+          file: "{{ forge_container_file | default('Containerfile') }}"
+          format: "{{ forge_container_format | default('oci') }}"
         pull: false
         push: true
         push_args:
